guest user, cookies

+++ store/models.py
@@ -14,7 +14,7 @@ class Customer(models.Model):  # customer model
 
 class Product(models.Model):
     name = models.CharField(max_length=200)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=7, decimal_places=2)
     digital = models.BooleanField(
         default=False, null=True, blank=True)  # doesn't require shipment
     image = models.ImageField(null=True, blank=True)
